Simple Vehicle management system
The aim of this project is to create a simple Vehicle management system where users can choose a
type of vehicle, decorate it with optional features, select a driving strategy and optionally add an
observer (like a monitoring system). In this project our aim was to understand and apply: Factory ,
Observer, Decorator and Strategy design patterns in a more easy and understandable way.